+++ test/palette.test.js
@@ -52,7 +52,7 @@ describe('mapnik.Palette ', function() {
 
         // Test rendering a blank image
         var filename = helper.filename();
-        var buffer = map.renderSync('png', {palette: pal});
+        var buffer = map.renderSync({palette: pal});
         assert.ok(buffer.length < 7300);
         map.renderFileSync(filename, {palette: pal});
         var stat = fs.statSync(filename);